+++ x/finality/types/expected_keepers.go
@@ -32,7 +32,7 @@ type CheckpointingKeeper interface {
 // IncentiveKeeper defines the expected interface needed for distributing rewards
 // and refund transaction fee for finality signatures
 type IncentiveKeeper interface {
-	RewardBTCStaking(ctx context.Context, height uint64, filteredDc *VotingPowerDistCache)
+	RewardBTCStaking(ctx context.Context, height uint64, filteredDc *VotingPowerDistCache, voters map[string]struct{})
 	IndexRefundableMsg(ctx context.Context, msg sdk.Msg)
 	BtcDelegationActivated(ctx context.Context, fp, del sdk.AccAddress, sat uint64) error
 	BtcDelegationUnbonded(ctx context.Context, fp, del sdk.AccAddress, sat uint64) error