+++ CHANGELOG.md
@@ -63,6 +63,7 @@ for rewards
 
 - [#341](https://github.com/babylonlabs-io/babylon/pull/341) Select parameters
 for pre-approval flow based on BTC LC tip height
+- [#360](https://github.com/babylonlabs-io/babylon/pull/360) Refactor rewarding
 
 ## v0.18.2